(cleanup) renamed auth.js to authRouter.js
