+++ src/server/server.js
@@ -7,7 +7,7 @@ var cookieParser = require('cookie-parser');
 var path = require('path');
 
 var schema = require('../db/schema.js')
-var authRouter = require('./routers/auth.js');
+var authRouter = require('./routers/authRouter.js');
 var docRouter = require('./routers/docRouter.js');
 
 var port = process.env.PORT || 3000;
